+++ src/Opulence/Orm/Repositories/IRepository.php
@@ -21,7 +21,7 @@ interface IRepository
      * @param object $entity The entity to add
      * @throws OrmException Thrown if the entity couldn't be added
      */
-    public function add(&$entity);
+    public function add($entity);
 
     /**
      * Deletes an entity from the repo
@@ -29,7 +29,7 @@ interface IRepository
      * @param object $entity The entity to delete
      * @throws OrmException Thrown if the entity couldn't be deleted
      */
-    public function delete(&$entity);
+    public function delete($entity);
 
     /**
      * Gets all the entities